Banco de Dados/requirements.txt
psycopg2-binary
requests
beautifulsoup4
python-dotenv
pandas
tqdm